move six to lib dir
