+++ scripts/lib/CIME/BuildTools/macrowriterbase.py
@@ -14,7 +14,7 @@ import os
 from abc import ABCMeta, abstractmethod
 from CIME.XML.standard_module_setup import *
 from CIME.utils import get_cime_root
-from CIME.six import add_metaclass
+from six import add_metaclass
 
 logger = logging.getLogger(__name__)
